move functions from draw-map.js to added : utils/draw/path-{data,info,classes}.js
draw-map.js :
move allowPathsOutsideZoom to view-controls.js
add action closeToolTipA() : pathInfo.closeToolTip() in place of : actions.closeToolTipA = closeToolTip.
move to added utils/draw/path-classes.js :  pathColourDefault, use_path_colour_scale, path_colour_scale_domain_set, path_colour_scale,   showScaffoldFeatures (showScaffoldMarkers). showAsymmetricAliases is in flow-controls viewOptions,   setup of path_colour_{domain,scale} factored to form configurePathColour() and colouredFeaturesChanged(), featureScaffold, scaffolds, scaffoldFeatures, intervals, intervalNames, intervalTree, scaffoldTicks,  pathClassA(),    colouredAg(), classFromSet(), locationClasses(), pathClasses(), pathColourUpdate(), scaffoldLegendColourUpdate(),   clearScaffoldColours(),   getUsePatchColour() as getUsePathColour().
move to added utils/draw/path-data.js : pathDataIsLine, pathDataInG,  showHoverExtraText, tracedAxisScale,  featureNameOfData(), data_text(),     blockO(), featureLine2(), inside(), pointSegment(), featureLineS2(), featureLineS3(), featureLineS(), lineHoriz(), featureLine(), path(), pathU(), pathUg(), pathAliasGroup(), inRangeI(), inRangeI2(), featureAliasesText(), pathFeatureStore(), featureNameInRange(), featureInRange(), patham(), axisFeatureTick(), patham2(), featureY_(), featureY2(), log_path_data(), pathUpdate_(), log_foreground_g(), pathUpdate(), dataOfPath(), featureNameOfPath().
move to added utils/draw/path-info.js : toolTip, toolTipHovered, setupToolTip() factored from : Setup the path hover tool tip, setupMouseHover(), toolTipMouseOver(), toolTipMouseOut(), closeToolTip(), setupToolTipMouseHover(), handleMouseOver(), hidePathHoverToolTip(), handleMouseOut(),
move to added utils/draw/feature-info.js : showTickLocations(), splitting out showTickLocationsOfAxis().
move to hover.js : configureHorizTickHover() as configureHorizTickHover_orig.
discard initial test data for axis-tracks : oa.tracks[].

block-adj.js : draw() : use pathInfo in place of axisApi for setupMouseHover().
path-data.js : pathU() : use pathDataUtils in place of axisApi for patham().
path-info.js : handleMouse{Over,Out}() : add 1st param pathInfo and curry(...)(this).

declare Promise (doesn't solve jshint warning) in : paths-progressive.js, block-values.js, groups.js, d3-svg.js, upload-table.js.
axis-draw.js : use .stacksView.stacks in preference to .stacks, which it replaces.

axisBrush.js :
use pathDataUtils in place of axisApi for pathUpdate() in : resetZoom(), axisScaleChanged(), draw_flipRegion().
handleFeatureCircleMouseOver() : use pathInfo.getToolTip() in place of oa.toolTip.

synteny-blocks-draw.js : blockLine() : use pathDataUtils to access patham2().
zoomPanCalcs.js : add comment line from draw-map.js : inRange().
stacks.js : use Ember_set in place of Ember.set, to avoid using Ember global.
