+++ frontend/app/components/draw/path-data.js
@@ -17,6 +17,8 @@ import {
   AxisChrName,
  } from '../../utils/utility-chromosome';
 
+import { PathDataUtils } from '../../utils/draw/path-data';
+
 //------------------------------------------------------------------------------
 
 /* global d3 */
@@ -25,8 +27,6 @@ import {
 
 /** Used for CSS selectors targeting <g> and <path>-s generated by this component. */
 const className = "block-feature-path";
-
-let axisApi;
 
 let trace_path = 1;
 
@@ -79,20 +79,21 @@ export default Component.extend({
   pathU : /*Ember.computed('feature0', 'feature1', 'block0', 'block1',*/ function() {
     let pathData = this;
     // based on draw-map.js : pathU(), pathUg(); this is equivalent and can replace those functions. (related : dataOfPath()).
-    if (! axisApi)
-      axisApi = stacks.oa.axisApi;
+
     let p = [];
     let block0 = pathData.get('blockId0'),
     block1 = pathData.get('blockId1');
       // pathData.get('block0'), pathData.get('block1');
-    p[0] = axisApi.patham(block0, block1, pathData.get('feature0.name'), pathData.get('feature1.name'));
-    const axisChrName = AxisChrName(stacks.oa);
-    let axisName2MapChr = axisChrName.axisName2MapChr;
-    if (trace_path > 1)
+    const pathDataUtils = PathDataUtils(stacks.oa);
+    p[0] = pathDataUtils.patham(block0, block1, pathData.get('feature0.name'), pathData.get('feature1.name'));
+    if (trace_path > 1) {
+      const axisChrName = AxisChrName(stacks.oa);
+      let axisName2MapChr = axisChrName.axisName2MapChr;
       console.log(
         "pathU",
         axisName2MapChr(block0), axisName2MapChr(block1),
         pathData.feature0, pathData.feature1, p[0]);
+    }
     return p;
   }/*)*/,
