+++ frontend/app/templates/components/panel/left-panel.hbs
@@ -67,7 +67,7 @@
                 pathColourScale=pathColourScale
                 scaffolds=scaffolds
                 scaffoldMarkers=scaffoldMarkers
-                showScaffoldMarkers=showScaffoldMarkers
+                showScaffoldFeatures=showScaffoldFeatures
                 showAsymmetricAliases=showAsymmetricAliases
                 selectedFeatures=selectedFeatures
                 updateFeaturesInBlocks="updateFeaturesInBlocks"
