+++ frontend/app/templates/mapview.hbs
@@ -19,7 +19,7 @@
     pathColourScale=pathColourScale
     scaffolds=scaffolds
     scaffoldMarkers=scaffoldMarkers
-    showScaffoldMarkers=showScaffoldMarkers
+    showScaffoldFeatures=showScaffoldFeatures
     showAsymmetricAliases=showAsymmetricAliases
     selectedFeatures=selectedFeatures
     updateFeaturesInBlocks="updateFeaturesInBlocks"
@@ -60,7 +60,7 @@
         pathColourScale=pathColourScale
         scaffolds=scaffolds
         scaffoldMarkers=scaffoldMarkers
-        showScaffoldMarkers=showScaffoldMarkers
+        showScaffoldFeatures=showScaffoldFeatures
         showAsymmetricAliases=showAsymmetricAliases
         highlightFeature=model.params.highlightFeature
         params=model.params
@@ -172,7 +172,7 @@
             data=selectedFeatures
             scaffolds=scaffolds
             scaffoldMarkers=scaffoldMarkers
-            showScaffoldMarkers=showScaffoldMarkers}}
+            showScaffoldFeatures=showScaffoldFeatures}}
 
         {{/if}}
