+++ frontend/app/utils/draw/d3-svg.js
@@ -2,6 +2,12 @@ import { later as run_later } from '@ember/runloop';
 
 /*----------------------------------------------------------------------------*/
 
+/* global Promise */
+
+//------------------------------------------------------------------------------
+
+const dLog = console.debug;
+
 const trace = 0;
 
 /*----------------------------------------------------------------------------*/
